% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FizzBuzz.R
\name{FizzBuzz}
\alias{FizzBuzz}
\title{Create a vector or FizzBuzz responses.}
\usage{
FizzBuzz(x)
}
\arguments{
\item{x}{A vector of numbers}
}
\value{
A vector with FizzBuzz responses
}
\description{
Takes a vector of numbers and returns a vector of FizzBuzz responses, where
numbers that are evenly divisible by 3 are substituted with "Fizz", numbers
that are divisible by 5 are replaced with "Buzz", and numbers divisible by
both 3 and 5 are substituted with "FizzBuzz". Negative numbers, zero, and
infinite values will return an error.
}
\examples{
FizzBuzz(seq(1,15,1))
FizzBuzz(c(1,2,3,5,11,30))
}
