% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Flagstaff_Weather.R
\docType{data}
\name{Flagstaff_Weather}
\alias{Flagstaff_Weather}
\title{Weather information at Flagstaff’s Pulliam Airport from 1950 to 2019.}
\format{
a data frame with 22967 observations
\describe{
  \item{DATE}{The date of observation as a POSIX date format.}
  \item{PRECIP}{Daily precipitation given in inches}
  \item{SNOW}{Daily snow given in inches}
  \item{TMAX}{Daily maximum temperature in degrees Farhenheit.}
  \item{TMIN}{Daily minimum temperature in degrees Farhenheit.}
}
}
\source{
\url{https://www.ncdc.noaa.gov/cdo-web/search}
}
\usage{
Flagstaff_Weather
}
\description{
Weather information at Flagstaff’s Pulliam Airport from 1950 to 2019.
}
\keyword{datasets}
